Add useScrollIntoView hook
