+++ extensions/ql-vscode/src/view/results/RawTableRow.tsx
@@ -2,7 +2,6 @@ import * as React from "react";
 import { ResultRow } from "../../common/bqrs-cli-types";
 import { selectedRowClassName, zebraStripe } from "./result-table-utils";
 import RawTableValue from "./RawTableValue";
-import { ScrollIntoViewHelper } from "./scroll-into-view-helper";
 
 interface Props {
   rowIndex: number;
@@ -10,8 +9,8 @@ interface Props {
   databaseUri: string;
   className?: string;
   selectedColumn?: number;
+  selectedItemRef?: React.Ref<any>;
   onSelected?: (row: number, column: number) => void;
-  scroller?: ScrollIntoViewHelper;
 }
 
 export default function RawTableRow(props: Props) {
@@ -26,7 +25,7 @@ export default function RawTableRow(props: Props) {
         const isSelected = props.selectedColumn === columnIndex;
         return (
           <td
-            ref={props.scroller?.ref(isSelected)}
+            ref={isSelected ? props.selectedItemRef : undefined}
             key={columnIndex}
             {...(isSelected ? { className: selectedRowClassName } : {})}
           >
